pyproject.toml
[tool.poetry]
name = "cgm"
version = "0.1.0"
description = "Captain Grant Manager"
authors = ["Paul Glushak <paul.glushak@forter.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
